mv to worker_test.go
