+++ miner/algo_greedy_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/txpool"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/ofac"
 	"github.com/holiman/uint256"
 )
 
@@ -112,89 +111,6 @@ func TestTxWithMinerFeeHeap(t *testing.T) {
 			orders.Shift()
 		} else if order.Bundle() != nil {
 			orders.Pop()
-		}
-	}
-}
-
-func TestGreedyComplianceList(t *testing.T) {
-	algos := []AlgoType{ALGO_GREEDY, ALGO_GREEDY_BUCKETS, ALGO_GREEDY_MULTISNAP, ALGO_GREEDY_BUCKETS_MULTISNAP}
-	expectations := make(map[string]int)
-	expectations["empty"] = 4
-	expectations["blacklistSender"] = 3
-	expectations["blacklistTo"] = 2
-	expectations["blacklist0xff"] = 3
-
-	for complianceListName, txCount := range expectations {
-		for _, algo := range algos {
-			//sender, to should be checked in the compliance list
-			statedb, chData, signers := genTestSetup(GasLimit)
-			blacklistedSigner := signers.addresses[1]
-			tx1 := signers.signTx(1, 21000, big.NewInt(1), big.NewInt(1), signers.addresses[5], big.NewInt(0), []byte{})
-			tx2 := signers.signTx(2, 21000, big.NewInt(1), big.NewInt(1), signers.addresses[6], big.NewInt(0), []byte{})
-			tx3 := signers.signTx(3, 21000, big.NewInt(1), big.NewInt(1), signers.addresses[7], big.NewInt(0), []byte{})
-			tx4 := signers.signCreate(4, 100000, big.NewInt(1), big.NewInt(1), big.NewInt(100), contractSend0xff)
-			txs := make(map[common.Address][]*txpool.LazyTransaction)
-			txs[blacklistedSigner] = []*txpool.LazyTransaction{{
-				Hash:      tx1.Hash(),
-				Tx:        tx1,
-				Time:      tx1.Time(),
-				GasFeeCap: uint256.MustFromBig(tx1.GasFeeCap()),
-				GasTipCap: uint256.MustFromBig(tx1.GasTipCap()),
-				GasPrice:  uint256.MustFromBig(tx1.GasPrice()),
-			}}
-			txs[signers.addresses[2]] = []*txpool.LazyTransaction{{
-				Hash:      tx2.Hash(),
-				Tx:        tx2,
-				Time:      tx2.Time(),
-				GasFeeCap: uint256.MustFromBig(tx2.GasFeeCap()),
-				GasTipCap: uint256.MustFromBig(tx2.GasTipCap()),
-				GasPrice:  uint256.MustFromBig(tx2.GasPrice()),
-			}}
-			txs[signers.addresses[3]] = []*txpool.LazyTransaction{{
-				Hash:      tx3.Hash(),
-				Tx:        tx3,
-				Time:      tx3.Time(),
-				GasFeeCap: uint256.MustFromBig(tx3.GasFeeCap()),
-				GasTipCap: uint256.MustFromBig(tx3.GasTipCap()),
-				GasPrice:  uint256.MustFromBig(tx3.GasPrice()),
-			}}
-			txs[signers.addresses[4]] = []*txpool.LazyTransaction{{
-				Hash:      tx4.Hash(),
-				Tx:        tx4,
-				Time:      tx4.Time(),
-				GasFeeCap: uint256.MustFromBig(tx4.GasFeeCap()),
-				GasTipCap: uint256.MustFromBig(tx4.GasTipCap()),
-				GasPrice:  uint256.MustFromBig(tx4.GasPrice()),
-			}}
-
-			ofac.UpdateComplianceLists(map[string]ofac.ComplianceList{
-				"empty":           {},
-				"blacklistSender": {blacklistedSigner: {}},
-				"blacklistTo":     {*tx2.To(): {}, *tx3.To(): {}},
-				"blacklist0xff":   {common.HexToAddress("0xff"): {}},
-			})
-
-			env := newEnvironment(chData, statedb, signers.addresses[0], 200000, big.NewInt(1))
-
-			var result *environment
-			switch algo {
-			case ALGO_GREEDY:
-				builder := newGreedyBuilder(chData.chain, chData.chainConfig, &defaultAlgorithmConfig, complianceListName, env, nil, nil)
-				result, _, _ = builder.buildBlock([]types.SimulatedBundle{}, nil, txs)
-			case ALGO_GREEDY_MULTISNAP:
-				builder := newGreedyMultiSnapBuilder(chData.chain, chData.chainConfig, &defaultAlgorithmConfig, complianceListName, env, nil, nil)
-				result, _, _ = builder.buildBlock([]types.SimulatedBundle{}, nil, txs)
-
-			case ALGO_GREEDY_BUCKETS:
-				builder := newGreedyBucketsBuilder(chData.chain, chData.chainConfig, &defaultAlgorithmConfig, complianceListName, env, nil, nil)
-				result, _, _ = builder.buildBlock([]types.SimulatedBundle{}, nil, txs)
-			case ALGO_GREEDY_BUCKETS_MULTISNAP:
-				builder := newGreedyBucketsMultiSnapBuilder(chData.chain, chData.chainConfig, &defaultAlgorithmConfig, complianceListName, env, nil, nil)
-				result, _, _ = builder.buildBlock([]types.SimulatedBundle{}, nil, txs)
-			}
-			if result.tcount != txCount {
-				t.Fatalf("Incorrect tx count [found: %d] expected %d, algoType : %v, compliance list %v", result.tcount, txCount, algo, complianceListName)
-			}
 		}
 	}
 }
